Replace lookup-controller util with Route#controllerFor
The logic in `/utils/lookup-controller.js` seems to be roughly the same that is
already implemented in [Route#controllerFor](https://emberjs.com/api/ember/2.15/classes/Ember.Route/methods/controllerFor?anchor=controllerFor).

Additionally, this seems to fix a bug that I was running into when
trying to use this addon inside of an engine with `factoryFor` not being
defined.