fix: remove empty default hooks from `QueryParams`
Because of the order that the `queryParams` mixin is applied, `queryParamsDidChange`, `setup`, and `reset` are not overrideable by methods in the controller.

The issue is visible on the [demo page](https://offirgolan.github.io/ember-parachute/); the progress bar that is supposed to appear on top of the form on `queryParamsDidChange` never shows up.

Removing the empty placeholder methods allows them to be defined in the controller, and they still work without being overriden because they are always called with `tryInvoke`.